[REF-3] changes: - compiling errors fix

+++ internal/model/search_options.go
@@ -56,6 +56,6 @@ type SearchOptions struct {
 	Fields []string `json:"fields,omitempty"`
 }
 
-func (s *SearchOptions) GetSize() int32 {
-	return int32(s.Size)
+func (s *SearchOptions) GetSize() int {
+	return s.Size
 }
